requirements.txt
faiss-cpu==1.7.4
langchain==0.0.261
openai==0.27.6
streamlit==1.25.0
tiktoken==0.3.1
tqdm==4.65.0
importlib==1.0.4
importlib-metadata==4.6.4
